+++ lib/src/iterator.dart
@@ -7,39 +7,41 @@ import 'utils.dart';
 
 /// Class that used to read data from [Pickle] objects.
 class PickleIterator {
+  /// Index where current chunk of data starts.
+  int readIndex = 0;
+
   /// Data stored in [Pickle] attached to this [PickleIterator].
-  late Uint8List payload;
+  @Deprecated("Use Pickle.header instead")
+  Uint8List get payload => pickle.header;
 
   /// Offset of actual [payload] data.
-  late int payloadOffset;
-
-  /// Index where current chunk of data starts.
-  late int readIndex;
+  @Deprecated("Use Pickle.headerSize instead")
+  int get payloadOffset => pickle.headerSize;
 
   /// Index where current chuck of data ends.
-  late int endIndex;
+  @Deprecated("Use Pickle.payloadSize instead")
+  int get endIndex => pickle.payloadSize;
+
+
+  /// [Pickle] attached to this [PickleIterator].
+  Pickle pickle;
 
   /// Creates new [PickleIterator] instance that can be used to read data from [Pickle] objects.
-  PickleIterator(Pickle pickle) {
-    payload = pickle.header;
-    payloadOffset = pickle.headerSize;
-    readIndex = 0;
-    endIndex = pickle.getPayloadSize();
-  }
+  PickleIterator(this.pickle);
 
   /// Returns the `bytes` value of the [Pickle] object at the specified buffer [length].
   dynamic readBytes<ReturnType>(int length,
       [ReturnType Function(int)? method]) {
     int readPayloadOffset = getReadPayloadOffsetAndAdvance(length);
     if (method != null) return method(readPayloadOffset);
-    return payload.sublist(readPayloadOffset, readPayloadOffset + length);
+    return pickle.header.sublist(readPayloadOffset, readPayloadOffset + length);
   }
 
   /// Returns current [Pickle] object value as `int32` and seeks to next data. A [RangeError] exception would be thrown when failed.
   int readInt() {
     return readBytes<int>(
         PickleSize.int32.value,
-        (readPayloadOffset) => ByteData.view(payload.buffer)
+        (readPayloadOffset) => pickle.headerView
             .getInt32(readPayloadOffset, Endian.little));
   }
 
@@ -47,7 +49,7 @@ class PickleIterator {
   int readUInt32() {
     return readBytes<int>(
         PickleSize.uint32.value,
-        (readPayloadOffset) => ByteData.view(payload.buffer)
+        (readPayloadOffset) => pickle.headerView
             .getUint32(readPayloadOffset, Endian.little));
   }
 
@@ -55,7 +57,7 @@ class PickleIterator {
   int readInt64() {
     return readBytes<int>(
         PickleSize.int64.value,
-        (readPayloadOffset) => ByteData.view(payload.buffer)
+        (readPayloadOffset) => pickle.headerView
             .getInt64(readPayloadOffset, Endian.little));
   }
 
@@ -63,7 +65,7 @@ class PickleIterator {
   int readUInt64() {
     return readBytes<int>(
         PickleSize.uint64.value,
-        (readPayloadOffset) => ByteData.view(payload.buffer)
+        (readPayloadOffset) => pickle.headerView
             .getUint64(readPayloadOffset, Endian.little));
   }
 
@@ -71,7 +73,7 @@ class PickleIterator {
   double readFloat() {
     return readBytes<double>(
         PickleSize.float.value,
-        (readPayloadOffset) => ByteData.view(payload.buffer)
+        (readPayloadOffset) => pickle.headerView
             .getFloat32(readPayloadOffset, Endian.little));
   }
 
@@ -79,7 +81,7 @@ class PickleIterator {
   double readDouble() {
     return readBytes<double>(
         PickleSize.double.value,
-        (readPayloadOffset) => ByteData.view(payload.buffer)
+        (readPayloadOffset) => pickle.headerView
             .getFloat64(readPayloadOffset, Endian.little));
   }
 
@@ -95,11 +97,11 @@ class PickleIterator {
 
   /// Gets read payload offset and seeks for next data in current [Pickle] object.
   int getReadPayloadOffsetAndAdvance(int length) {
-    if (length > endIndex - readIndex) {
-      readIndex = endIndex;
+    if (length > pickle.payloadSize - readIndex) {
+      readIndex = pickle.payloadSize;
       throw RangeError('Failed to read data with length of $length');
     }
-    int readPayloadOffset = payloadOffset + readIndex;
+    int readPayloadOffset = pickle.headerSize + readIndex;
     advance(length);
     return readPayloadOffset;
   }
@@ -107,8 +109,8 @@ class PickleIterator {
   /// Seeks for next data in current [Pickle] object.
   void advance(int size) {
     int alignedSize = alignInt(size, PickleSize.uint32.value);
-    if (endIndex - readIndex < alignedSize) {
-      readIndex = endIndex;
+    if (pickle.payloadSize - readIndex < alignedSize) {
+      readIndex = pickle.payloadSize;
     } else {
       readIndex += alignedSize;
     }